backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
python-dotenv==1.1.1
openai==1.99.9
chromadb==1.0.15
faiss-cpu==1.12.0
presidio-analyzer==2.2.357
presidio-anonymizer==2.2.357
llmlingua==0.2.2
langsmith
